test: compare keyless sandbox observer usage field-by-field
