+++ crates/mega-evm/tests/rex2/main.rs
@@ -2,5 +2,6 @@
 
 mod keyless_deploy;
 mod keyless_sandbox_observer;
+mod keyless_sandbox_support;
 mod oracle_hint;
 mod selfdestruct;
